# packages/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: SRapi API
  version: 0.1.0
  description: |
    This is the first executable OpenAPI contract for SRapi M1.
    It intentionally covers only the MVP bootstrap surface so the project can
    generate server types, TypeScript clients, contract tests, and mock servers
    before implementation starts.
  license:
    name: Proprietary
    url: https://srapi.local/license
servers:
  - url: /
    description: Relative API root. Local development usually serves this from http://localhost:8080.
...
paths:
...
  /api/v1/auth/login:
    post:
      tags:
        - Auth
      operationId: login
      summary: Login to the console.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login succeeded and session cookie was set.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIdHeader"
            Set-Cookie:
              description: HttpOnly console session cookie.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/StandardError"
  /api/v1/auth/logout:
    post:
      tags:
        - Auth
      operationId: logout
      summary: Logout from the console.
      security:
        - cookieAuth: []
          csrfHeader: []
      responses:
        "204":
          description: Logout succeeded.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIdHeader"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/StandardError"
  /api/v1/me:
    get:
      tags:
        - CurrentUser
      operationId: getCurrentUser
      summary: Get current console user.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Current user.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIdHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/StandardError"
  /api/v1/me/usage:
    get:
      tags:
        - CurrentUser
      operationId: getCurrentUserUsage
      summary: List usage logs for the current console user.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Current user usage logs.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIdHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageLogListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/StandardError"
  /api/v1/me/subscriptions:
    get:
      tags:
        - CurrentUser
      operationId: getCurrentUserSubscriptions
      summary: List subscriptions for the current console user.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Current user subscriptions.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIdHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSubscriptionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/StandardError"
  /api/v1/payment/methods:
    get:
      tags:
        - Payment
      operationId: listPaymentMethods
      summary: List available payment methods for the current user.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Available payment methods.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIdHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethodListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/StandardError"
  /api/v1/payment/orders:
    get:
      tags:
        - Payment
      operationId: listPaymentOrders
      summary: List payment orders for the current user.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Current user payment orders.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIdHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrderListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/StandardError"
    post:
      tags:
        - Payment
      operationId: createPaymentOrder
      summary: Create a payment order for the current user.
      security:
        - cookieAuth: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentOrderRequest"
      responses:
        "201":
          description: Payment order created.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIdHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrderResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/StandardError"
  /api/v1/payment/orders/{id}:
    get:
      tags:
        - Payment
      operationId: getPaymentOrder
      summary: Get a payment order owned by the current user.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Id"
      responses:
        "200":
          description: Payment order.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIdHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrderResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/StandardError"
        default:
          $ref: "#/components/responses/StandardError"
  /api/v1/payment/orders/{id}/cancel:
    post:
      tags:
        - Payment
      operationId: cancelPaymentOrder
      summary: Cancel a pending payment order owned by the current user.
      security:
        - cookieAuth: []
          csrfHeader: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Id"
      responses:
        "200":
          description: Payment order canceled.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIdHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrderResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/StandardError"
  /api/v1/webhooks/payments/{provider}:
    post:
      tags:
        - Payment
      operationId: handlePaymentWebhook
      summary: Handle a signed payment provider webhook.
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentWebhookRequest"
      responses:
        "200":
          description: Webhook was accepted or already handled.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIdHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentWebhookResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/StandardError"
  /api/v1/api-keys:
    get:
      tags:
        - ApiKeys
      operationId: listApiKeys
      summary: List API keys visible to the current user.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Status"
      responses:
        "200":
          description: API key list.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIdHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/StandardError"
    post:
      tags:
        - ApiKeys
      operationId: createApiKey
      summary: Create an API key and return its plaintext exactly once.
      security:
        - cookieAuth: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "201":
          description: API key created. The plaintext key is only returned in this response.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIdHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateApiKeyResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/StandardError"
...
  /api/v1/admin/overview:
    get:
      tags:
        - AdminScheduler
      operationId: getAdminOverview
      summary: Get admin overview counters.
      security:
        - cookieAuth: []
      x-srapi-rbac:
        owner: read
        admin: read
        operator: read
      responses:
        "200":
          description: Admin overview.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestIdHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOverviewResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/StandardError"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: srapi_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
    gatewayBearerAuth:
      type: http
      scheme: bearer
  headers:
    RequestIdHeader:
      description: Request correlation id.
      schema:
        $ref: "#/components/schemas/RequestId"
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
...
  responses:
...
    ValidationError:
      description: Request validation failed.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestIdHeader"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validation:
              value:
                error:
                  code: VALIDATION_FAILED
                  message: validation failed
                  details: {}
                  trace_id: trace_01HY0000000000000000000000
                request_id: req_01HY0000000000000000000000
...
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
...
    LoginResponse:
      type: object
      required:
        - data
        - request_id
      properties:
        data:
          $ref: "#/components/schemas/SessionData"
        request_id:
          $ref: "#/components/schemas/RequestId"
...